Got rid of confirming store state
