+++ app/shared/components/Wallet/Panel/Form/Stake.js
@@ -1,14 +1,16 @@
 // @flow
 import React, { Component } from 'react';
-import { Segment } from 'semantic-ui-react';
+import { Decimal } from 'decimal.js';
+import { Segment, Form, Divider, Message, Button } from 'semantic-ui-react';
 
 import WalletPanelFormStakeStats from './Stake/Stats';
-import WalletPanelFormStakeInputs from './Stake/Inputs';
+import WalletPanelFormStakeInput from './Stake/Input';
+
 import FormMessageTransactionSuccess from '../../../Global/Form/Message/TransactionSuccess';
+import WalletPanelFormStakeConfirming from './Stake/Confirming';
+
 
 import FormMessageError from '../../../Global/Form/Message/Error';
-
-import { Decimal } from 'decimal.js';
 
 type Props = {
   actions: {},
@@ -33,11 +35,72 @@ export default class WalletPanelFormStake extends Component<Props> {
     const parsed_net_weight = net_weight.split(' ')[0];
 
     this.state = {
-      cpuAmount: Decimal(parsed_cpu_weight),
+      decimalCpuAmount: Decimal(parsed_cpu_weight),
       cpuOriginal: Decimal(parsed_cpu_weight),
-      netAmount: Decimal(parsed_net_weight),
-      netOriginal: Decimal(parsed_net_weight)
+      decimalNetAmount: Decimal(parsed_net_weight),
+      netOriginal: Decimal(parsed_net_weight),
+      confirming: false,
+      formError: null,
+      submitDisabled: true
     };
+  }
+
+  onSubmit = (e) => {
+    if (!this.state.submitDisabled) {
+      this.setState({
+        confirming: true
+      });
+    }
+    e.preventDefault();
+    return false;
+  }
+
+  onKeyPress = (e) => {
+    if (e.key === 'Enter') {
+      this.onSubmit(e);
+
+      e.preventDefault();
+      return false;
+    }
+  }
+
+  onError = (errorMessage) => {
+    this.setState({
+      submitDisabled: true,
+      formError: errorMessage
+    });
+  }
+
+  onChange = (name, value) => {
+    const decimalFieldName = `decimal${name.charAt(0).toUpperCase()}${name.slice(1)}`;
+
+    this.setState({
+      submitDisabled: false,
+      formError: null,
+      [decimalFieldName]: Decimal(value)
+    });
+  }
+
+  onConfirm = () => {
+    const {
+      account,
+      actions
+    } = this.props;
+
+    const {
+      cpuAmount,
+      netAmount
+    } = this.state;
+
+    const {
+      setStake
+    } = actions;
+
+    this.setState({
+      confirming: false
+    });
+
+    setStake(account, netAmount, cpuAmount);
   }
 
   render() {
@@ -47,14 +110,16 @@ export default class WalletPanelFormStake extends Component<Props> {
       balance,
       onClose,
       system,
-      validate
+      validate,
+      t
     } = this.props;
 
     const {
-      cpuAmount,
+      decimalCpuAmount,
       cpuOriginal,
-      netAmount,
-      netOriginal
+      decimalNetAmount,
+      netOriginal,
+      submitDisabled
     } = this.state;
 
     const EOSbalance = balance.EOS || 0;
@@ -75,9 +140,10 @@ export default class WalletPanelFormStake extends Component<Props> {
       lastTransactions.push(system.UNDELEGATEBW_LAST_TRANSACTION);
     }
 
-    const shouldShowSuccess = (system.DELEGATEBW === 'SUCCESS' || system.UNDELEGATEBW === 'SUCCESS');
-    const shouldShowError = (system.DELEGATEBW === 'FAILURE' || system.UNDELEGATEBW === 'FAILURE');
-    const shouldShowForm = (validate.STAKE === 'NULL' || validate.STAKE === 'CONFIRMING') && !shouldShowSuccess && !shouldShowError;
+    const shouldShowConfirm = this.state.confirming;
+    const shouldShowSuccess = (system.DELEGATEBW === 'SUCCESS' || system.UNDELEGATEBW === 'SUCCESS') && !shouldShowConfirm;
+    const shouldShowError = (system.DELEGATEBW === 'FAILURE' || system.UNDELEGATEBW === 'FAILURE') && !shouldShowConfirm;
+    const shouldShowForm = (validate.STAKE === 'NULL') && !shouldShowSuccess && !shouldShowError && !shouldShowConfirm;
 
     return (
       <Segment
@@ -102,17 +168,63 @@ export default class WalletPanelFormStake extends Component<Props> {
                 EOSbalance={EOSbalance}
                 netOriginal={netOriginal}
               />
-              <WalletPanelFormStakeInputs
-                actions={actions}
-                account={account}
-                cpuAmount={cpuAmount}
-                cpuOriginal={cpuOriginal}
-                EOSbalance={EOSbalance}
-                netAmount={netAmount}
-                netOriginal={netOriginal}
-                onClose={onClose}
-                validate={validate}
-              />
+              <Form
+                onKeyPress={this.onKeyPress}
+                onSubmit={this.onSubmit}
+              >
+                <Form.Group widths="equal">
+                  <WalletPanelFormStakeInput
+                    cpuOriginal={cpuOriginal}
+                    decimalCpuAmount={decimalCpuAmount}
+                    decimalNetAmount={decimalNetAmount}
+                    defaultValue={decimalCpuAmount}
+                    EOSbalance={EOSbalance}
+                    label={t('update_staked_cpu_amount')}
+                    name='cpuAmount'
+                    netOriginal={netOriginal}
+                    onChange={this.onChange}
+                    onError={this.onError}
+                  />
+                  <WalletPanelFormStakeInput
+                    cpuOriginal={cpuOriginal}
+                    decimalCpuAmount={decimalCpuAmount}
+                    decimalNetAmount={decimalNetAmount}
+                    defaultValue={decimalNetAmount}
+                    EOSbalance={EOSbalance}
+                    label={t('update_staked_net_amount')}
+                    name='netAmount'
+                    netOriginal={netOriginal}
+                    onChange={this.onChange}
+                    onError={this.onError}
+                  />
+                </Form.Group>
+                {(this.state.formError)
+                  ? (
+                    <Message
+                      failure
+                      content={t(this.state.formError)}
+                    />
+                  ) : ''}
+                <Divider />
+                <Message
+                  icon="info circle"
+                  info
+                  content={t('undelegate_explanation')}
+                />
+                <Divider />
+                <Button
+                  content={t('cancel')}
+                  color="grey"
+                  onClick={onClose}
+                />
+                <Button
+                  content={t('update_staked_coins')}
+                  color="green"
+                  disabled={submitDisabled}
+                  floated="right"
+                  primary
+                />
+              </Form>
             </div>
           ) : ''}
 
@@ -122,6 +234,22 @@ export default class WalletPanelFormStake extends Component<Props> {
               error={system.DELEGATEBW_LAST_ERROR ||
                      system.UNDELEGATEBW_LAST_ERROR}
               onClose={onClose}
+            />
+          ) : ''}
+        {(shouldShowConfirm)
+          ? (
+            <WalletPanelFormStakeConfirming
+              account={account}
+              actions={actions}
+              decimalCpuAmount={decimalCpuAmount}
+              cpuOriginal={cpuOriginal}
+              EOSbalance={EOSbalance}
+              decimalNetAmount={decimalNetAmount}
+              netOriginal={netOriginal}
+              onClose={onClose}
+              onConfirm={this.onConfirm}
+              validate={validate}
+              t={t}
             />
           ) : ''}
       </Segment>
